地图导入功能完成

+++ app/src/main/java/jmy/com/newsschool/util/MapCopyUtil.java
@@ -1,44 +1,58 @@
 package jmy.com.newsschool.util;
 
+import org.litepal.crud.DataSupport;
+
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
+import java.util.List;
+
+import jmy.com.newsschool.bean.CityData;
 
 public class MapCopyUtil {
-    public static int copyFiles(String fromDir,String toDir){
+    public static int copyFiles(String province, String fromDir, String toDir) {
         File fromFile = new File(fromDir);
-        if(!fromFile.exists()){
+        if (!fromFile.exists()) {
             return -1;//源文件夹不存在，检查u盘中是否有该文件夹
         }
         File toFile = new File(toDir);
-        if(!toFile.exists()){
+        if (!toFile.exists()) {
             toFile.mkdirs();
         }
         File[] listFile = fromFile.listFiles();
-        if(listFile==null||listFile.length==0){
+        if (listFile == null || listFile.length == 0) {
             return -2;//文件夹下没有文件
         }
-        for(File file :listFile){
-            if(!copyFile(file.getAbsolutePath(),toDir+"/"+file.getName())){
-                delete(toDir);
-                return 1;
+        for (File file : listFile) {
+            List<CityData> cityData1 = DataSupport.where("name = ?", file.getName()).find(CityData.class);
+            if (cityData1 != null&&cityData1.size()>0) {
+                continue;
             }
+            copyFile(file.getAbsolutePath(), toDir + "/" + file.getName());
+            CityData cityData = new CityData();
+            cityData.setName(file.getName());
+            cityData.setParentName(province);
+            cityData.save();
         }
         return 0;
     }
 
-    public static void delete(String toDir) {
+    public static void delete(String toDir, String proviceKey) {
         File toFile = new File(toDir);
-        if(!toFile.exists()){
+        if (!toFile.exists()) {
             return;
         }
-        for (File f : toFile.listFiles()){
-            f.delete();
+        List<CityData> citys = DataSupport.where("parentName = ?",proviceKey).find(CityData.class);
+        for(CityData c :citys){
+            File f = new File(toDir+"/"+c.getName());
+            if(f.exists()){
+                f.delete();
+            }
+            DataSupport.delete(CityData.class,c.getId());
         }
-        toFile.delete();
     }
 
     private static boolean copyFile(String fromFile, String toFile) {
